backend/FastApi/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
audioread==3.0.1
autopep8==2.3.2
beautifulsoup4==4.12.2
blinker==1.7.0
boto3==1.37.38
botocore==1.37.38
bs4==0.0.1
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cssselect==1.2.0
cycler==0.12.1
decorator==5.2.1
dnspython==2.7.0
email_validator==2.2.0
fake-useragent==1.2.1
fastapi==0.115.12
filelock==3.18.0
Flask==3.0.2
flask-mysql-connector==1.1.0
fonttools==4.48.1
fsspec==2025.3.2
h11==0.14.0
httptools==0.6.4
idna==3.4
imageio==2.34.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==4.9.3
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
msgpack==1.1.0
mysql-connector-python==8.3.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parse==1.19.1
pillow==10.2.0
platformdirs==4.3.7
pooch==1.8.2
psycopg==3.1.1
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
pyee==8.2.2
pyparsing==3.1.1
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
requests-html==0.10.0
Resemblyzer==0.1.4
s3transfer==0.11.5
scikit-learn==1.6.1
scipy==1.15.2
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.4.1
soxr==0.5.0.post1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
tqdm==4.66.1
typing==3.7.4.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
urllib3==1.26.16
uvicorn==0.34.2
w3lib==2.1.2
watchdog==6.0.0
watchfiles==1.0.5
webrtcvad==2.0.10
websockets==10.4
Werkzeug==3.0.1
zipp==3.16.2
